feat notas, curso.

+++ src/main/java/com/example/crudspring/service/EstudianteService.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class EstudianteService {
@@ -29,14 +30,32 @@ public class EstudianteService {
     }
 
     //Metodo para actualizar el estudiante
-    public Estudiante actualizarEstudiante(Estudiante estudiante){
-        return estudianteRepository.save(estudiante);
+    public Optional<Estudiante> actualizarEstudiante(Estudiante estudiante, Long id){
+        Optional<Estudiante> estudianteBuscado = estudianteRepository.findById(id);
+
+        if(estudianteBuscado.isPresent()){
+            Estudiante _estudiante = estudianteBuscado.get();
+            _estudiante.setNombre(estudiante.getNombre());
+            _estudiante.setApellido(estudiante.getApellido());
+            _estudiante.setCorreo(estudiante.getCorreo());
+            _estudiante.setUsuario(estudiante.getUsuario());
+            estudianteRepository.save(_estudiante);
+        }
+        return estudianteBuscado;
+
+
     }
 
     //Metodo para eliminar el estudiante
     public String eliminarEstudiante(Long id){
-        estudianteRepository.deleteById(id);
-        return "Estudiante eliminado correctamente";
+      Optional<Estudiante> estudianteBuscado = estudianteRepository.findById(id);
+        if (estudianteBuscado.isPresent()){
+            estudianteRepository.deleteById(id);
+            return "Estudiante eliminado correctamente";
+        }else{
+            return null;
+        }
+
     }
 
 }
